Redefine `PROPVARIANT`
In the case of the `PROPVARIANT` structure/union, it does not appear
to make much sense to keep the type declaration as similar as possible
to the original, because we would then need to define tons of type
aliases somehow, and many fields would be redundant (most pointer
fields would turn into `IntPtr` when only one is needed).

Instead, let's define one union field per type and name it accord-
ingly. This results in a shorter and cleaner structure, even if it
is not as complete as the original.