Remove `stakx` from names
This removes the `stakx.` prefix from namespaces, from file and folder
names, and from the repo URLs. This is to allow easier forking and to
not force people to see my name everywhere.

The name prefix does however stay in the package name, as that package
is the one branch published by me.